requirements.txt
# Core
pydantic>=2.0,<3.0
redis>=5.0,<6.0

# Agentes e orquestração
langgraph>=0.2,<1.0
anthropic>=0.30,<1.0

# Banco de dados
supabase>=2.0,<3.0
pgvector>=0.2,<1.0

# APIs externas
openai>=1.0,<2.0
elevenlabs>=1.0,<2.0
httpx>=0.27,<1.0

# Backend
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
websockets>=12.0,<14.0

# Utilitários
python-dotenv>=1.0,<2.0
pyyaml>=6.0,<7.0

# Testes
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
fakeredis>=2.23,<3.0
# Nota: board.py usa SET NX PX para locking (sem Lua scripts) — compatível com fakeredis
